Replacing Image/Map Palettes with toolbars (see changelog for details)
git-svn-id: file:///desktop/SourceForge%20Backup/sphere/svnrepos/trunk@897 c0eb2ce6-7f40-480a-a1e0-6c4aea08c2c2

+++ sphere/source/editor/SpritesetView.hpp
@@ -14,7 +14,7 @@ public:
   virtual void SV_SpritesetModified() = 0;
   virtual void SV_CopyCurrentFrame() = 0;
   virtual void SV_PasteCurrentFrame() = 0;
-	virtual void SV_ZoomFactorChanged(int zoom) = 0;
+	virtual void SV_ZoomFactorChanged(double zoom) = 0;
 };
 
 
@@ -29,8 +29,8 @@ public:
   void SetDirection(int direction);
   void SetFrame(int frame);
   
-  void SetZoomFactor(int zoom);
-  int GetZoomFactor() const;
+  void SetZoomFactor(double zoom);
+  double GetZoomFactor() const;
 
   void SpritesetResized();
 
@@ -80,7 +80,7 @@ private:
   
   CDIBSection* m_DrawBitmap;
 
-  int m_ZoomFactor;
+  double m_ZoomFactor;
   int m_MaxFrameWidth;
   int m_MaxFrameHeight;
